review: use uintptr_t to fix Windows build.

+++ test/extensions/common/wasm/test_data/test_cpp.cc
@@ -21,7 +21,7 @@ START_WASM_PLUGIN(CommonWasmTestCpp)
 
 // With Emscripten, dereferencing a null pointer does not immediately cause a segmentation fault,
 // so use an invalid address to trigger it.
-static int* badptr = reinterpret_cast<int*>(0xDEADBEEF);
+static uintptr_t* badptr = reinterpret_cast<uintptr_t*>(0xDEADBEEF);
 
 static float gNan = std::nan("1");
 static float gInfinity = INFINITY;
