+++ test/extensions/filters/network/wasm/test_data/test_cpp.cc
@@ -11,7 +11,7 @@ START_WASM_PLUGIN(NetworkTestCpp)
 
 // With Emscripten, dereferencing a null pointer does not immediately cause a segmentation fault,
 // so use an invalid address to trigger it.
-static int* badptr = reinterpret_cast<int*>(0xDEADBEEF);
+static uintptr_t* badptr = reinterpret_cast<uintptr_t*>(0xDEADBEEF);
 
 class ExampleContext : public Context {
 public:
